# restApi/openapi.json
{"openapi":"3.1.0","info":{"title":"Books API","description":"REST API для управления библиотекой книг","version":"1.0.0"},"paths":{"/":{"get":{"tags":["Root"],"summary":"Root","description":"Корневой эндпоинт API.\nВозвращает приветственное сообщение и ссылки на документацию.","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/books":{"get":{"tags":["Books"],"summary":"Get Books","description":"Получить список всех книг.\nВозвращает список всех книг, хранящихся в системе.","operationId":"get_books_api_books_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/Book"},"type":"array","title":"Response Get Books Api Books Get"}}}}}},"post":{"tags":["Books"],"summary":"Create Book","description":"Создать новую книгу.\nПринимает данные новой книги и добавляет её в систему.\nАвтоматически генерирует уникальный ID для книги.\nВозвращает созданную книгу с присвоенным ID.","operationId":"create_book_api_books_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Book"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Book"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/books/{book_id}":{"get":{"tags":["Books"],"summary":"Get Book","description":"Получить книгу по ID.\n- **book_id**: ID книги (целое число)\n\nВозвращает информацию о книге с указанным ID.\nЕсли книга не найдена, возвращается ошибка 404.","operationId":"get_book_api_books__book_id__get","parameters":[{"name":"book_id","in":"path","required":true,"schema":{"type":"integer","title":"Book Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Book"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["Books"],"summary":"Update Book","description":"Полностью обновить информацию о книге.","operationId":"update_book_api_books__book_id__put","parameters":[{"name":"book_id","in":"path","required":true,"schema":{"type":"integer","title":"Book Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Book"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Book"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"patch":{"tags":["Books"],"summary":"Partial Update Book","description":"Частично обновить информацию о книге.\n\n- **book_id**: ID книги для обновления\n- **book_update**: Данные для обновления (только указанные поля будут изменены)","operationId":"partial_update_book_api_books__book_id__patch","parameters":[{"name":"book_id","in":"path","required":true,"schema":{"type":"integer","title":"Book Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Book"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Books"],"summary":"Delete Book","description":"Удалить книгу по ID.\n\n- **book_id**: ID книги для удаления","operationId":"delete_book_api_books__book_id__delete","parameters":[{"name":"book_id","in":"path","required":true,"schema":{"type":"integer","title":"Book Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Book":{"properties":{"id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Id"},"title":{"type":"string","maxLength":200,"minLength":1,"title":"Title","description":"Название книги"},"author":{"type":"string","maxLength":100,"minLength":1,"title":"Author","description":"Автор книги"},"year":{"type":"integer","maximum":2026.0,"minimum":1000.0,"title":"Year","description":"Год издания"},"isbn":{"anyOf":[{"type":"string","maxLength":13,"minLength":10},{"type":"null"}],"title":"Isbn","description":"ISBN книги"}},"type":"object","required":["title","author","year"],"title":"Book","example":{"author":"Михаил Булгаков","isbn":"9785170123456","title":"Мастер и Маргарита","year":1967}},"BookUpdate":{"properties":{"title":{"anyOf":[{"type":"string","maxLength":200,"minLength":1},{"type":"null"}],"title":"Title"},"author":{"anyOf":[{"type":"string","maxLength":100,"minLength":1},{"type":"null"}],"title":"Author"},"year":{"anyOf":[{"type":"integer","maximum":2026.0,"minimum":1000.0},{"type":"null"}],"title":"Year"},"isbn":{"anyOf":[{"type":"string","maxLength":13,"minLength":10},{"type":"null"}],"title":"Isbn"}},"type":"object","title":"BookUpdate"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}